Make hasColors implementation reusable in tests
