+++ lib/worker/fake-tty.js
@@ -2,16 +2,12 @@
 const tty = require('tty');
 const ansiEscapes = require('ansi-escapes');
 const options = require('./options').get();
+const makeHasColors = require('./fake-tty-has-colors');
 
 const fakeTTYs = new Set();
 
 const {isatty} = tty;
 tty.isatty = fd => fakeTTYs.has(fd) || isatty(fd);
-
-// Call original method to ensure the correct errors are thrown.
-const assertHasColorsArguments = count => {
-	tty.WriteStream.prototype.hasColors(count);
-};
 
 const simulateTTY = (stream, {colorDepth, hasColors, columns, rows}) => {
 	Object.assign(stream, {isTTY: true, columns, rows});
@@ -42,15 +38,7 @@ const simulateTTY = (stream, {colorDepth, hasColors, columns, rows}) => {
 	}
 
 	if (hasColors) {
-		stream.hasColors = (count = 16, env) => {
-			// `count` is optional too, so make sure it's not an env object.
-			if (env === undefined && typeof count === 'object' && count !== null) {
-				count = 16;
-			}
-
-			assertHasColorsArguments(count);
-			return count <= 2 ** colorDepth;
-		};
+		stream.hasColors = makeHasColors(colorDepth);
 	}
 };
